refactor: move files into families
